requirements.txt
Babel
Django
Pillow
django-phonenumber-field
phonenumbers
psycopg2
reportlab